requirements.txt
fastapi
uvicorn[standard]
pillow
transformers
torch
torchvision
python-multipart
pytesseract
opencv-python